% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{add_BA_stats_geom_manual}
\alias{add_BA_stats_geom_manual}
\title{Manually add Bland Altman geometry to plot}
\usage{
add_BA_stats_geom_manual(
  bias,
  loa.lwr,
  loa.upr,
  exponentiated = FALSE,
  name_ref = "ref",
  name_alt = "alt"
)
}
\arguments{
\item{bias, loa.lwr, loa.upr}{estimates to be plotted. Optionally including confidence intervals
as `c(est, ci.lwr, ci.upr)`.}

\item{exponentiated}{Set true if estimates are exponentiated estimates from a model on log-transformed data.
Treats estimates as ratios.}

\item{name_ref, name_alt}{Name of reference and alternative method. Only used if `exponentiated = TRUE`}
}
\description{
Manually add Bland Altman geometry to plot
}
