% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{add_confint}
\alias{add_confint}
\title{Add confidence intervals to Bland-Altman analysis object.}
\usage{
add_confint(
  ba_obj,
  level = 0.95,
  nsim = 1999,
  .progress = "txt",
  PBargs = list(style = 3)
)
}
\arguments{
\item{ba_obj}{Bland-Altman analysis object}

\item{level}{Confidence level (default is 0.95)}

\item{nsim}{Number of bootstrap samples}

\item{.progress, PBargs}{see \code{\link[lme4:bootMer]{lme4::bootMer()}}}
}
\value{
Bland-Altman analysis object (\code{x}) with added confidence intervals
}
\description{
The confidence intervals are calculated with a percentile parametric bootstrap (see \code{\link[lme4:confint.merMod]{lme4::confint.merMod()}} for details).
}
\details{
Note that when calculating confidence intervals for percentage errors, the mean value is treated as a known value
rather than an estimate. Any uncertainty in the estimate of the mean is not included in the confidence intervals
for percentage errors. This is a technical limitation, since the bootstrap analysis is based on a single mixed model
(\code{diff ~ 1 + (1|id)}) which has no information about the mean value.
}
