% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{as.data.frame.ba_analysis}
\alias{as.data.frame.ba_analysis}
\title{Convert BA object to data.frame}
\usage{
\method{as.data.frame}{ba_analysis}(ba_obj)
}
\arguments{
\item{ba_obj}{}
}
\description{
Convert BA object to data.frame
}
