% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{BA_plot_normalized_log}
\alias{BA_plot_normalized_log}
\title{Plot relative Bland-Altman estimates on raw data.}
\usage{
BA_plot_normalized_log(
  ba_obj,
  show_subject_legend = FALSE,
  aspect_ratio = NULL
)
}
\arguments{
\item{ba_obj}{Bland-Altman analysis object}

\item{show_subject_legend}{Show legend for subjects}

\item{aspect_ratio}{Set aspect ratio (x/y) between X and Y axis (sets \code{ggplot2::coord_fixed()}), Default (NULL) is automatic.}
}
\value{
Bland-Altman style plot with relative differences plotted on absolute differences.
}
\description{
Visualizes a Bland-Altman analysis with log-transformed data (showing relative errors) on the non-transformed data.
}
\examples{
BA_CO <- compare_methods(CO, "ic", "rv", id_col = "sub", logtrans = TRUE)
BA_plot_normalized_log(BA_CO)

}
