% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CO}
\alias{CO}
\title{Cardiac output}
\format{
A data frame with 60 rows and 3 variables:
\describe{
\item{sub}{Subject}
\item{rv}{Cardiac output,  radionuclide ventriculography (RV)}
\item{ic}{Cardiac output,  impedance cardiography (IC)}
}
}
\source{
\url{https://www-users.york.ac.uk/~mb55/datasets/datasets.htm}
}
\usage{
CO
}
\description{
Cardiac output by two methods, RV and IS, for 12 subjects
}
\references{
\cite{Bland JM, Altman DG. (1999) Measuring agreement in method comparison studies. Statistical Methods in Medical Research 8, 135-160.}
(data provided by Dr LS Bowling)
}
\keyword{datasets}
