% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{compare_methods}
\alias{compare_methods}
\title{Compare methods}
\usage{
compare_methods(df, ref_col, alt_col, id_col)
}
\arguments{
\item{df}{Data frame}

\item{ref_col}{name of column containing reference measurements}

\item{alt_col}{name of column containing alternative measurements}

\item{id_col}{name of column containing unique subject id's}
}
\value{

}
\description{
Compare methods
}
\examples{
compare_methods(CO, ref_col = "rv", alt_col = "ic", id_col = "sub")
}
