% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{confint.ba_analysis}
\alias{confint.ba_analysis}
\title{Calculate confidence interval}
\usage{
\method{confint}{ba_analysis}(ba_obj, level = 0.95, nsim = 1999, .progress = "txt")
}
\arguments{
\item{ba_obj}{BA analysis object}

\item{level}{Confidence level (default is 0.95)}

\item{nsim}{Number of bootstrap samples}

\item{.progress}{see `?lme4::bootMer()`
``}
}
\value{
Matrix of bootstrap confidence intervals for BA statistics.
}
\description{
Calculate confidence interval
}
