% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{plot_BA}
\alias{plot_BA}
\title{Create Bland Altman Plot}
\usage{
plot_BA(
  ba_obj,
  aspect_ratio = NULL,
  show_subject_legend = FALSE,
  keep_log_scale = FALSE
)
}
\arguments{
\item{ba_obj}{BA analysis object}

\item{aspect_ratio}{Set aspect ratio (x/y) between X and Y axis (sets \code{coord_fixed()}), Default (NULL) is automatic.}

\item{show_subject_legend}{Show legend for subjects}

\item{keep_log_scale}{Show log transformed differences. If \code{FALSE} (default), values and parameters are exponentiated before plotting}
}
\value{
Bland Altman plot (ggplot)
}
\description{
Create Bland Altman Plot
}
\examples{
ba_obj <- compare_methods(CO, ic, rv, id_col = sub)
plot_BA(ba_obj)

}
