% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{plot_BA_complete}
\alias{plot_BA_complete}
\title{Plot all plots in extended Bland Altman analysis.}
\usage{
plot_BA_complete(
  ba_obj,
  show_subject_legend = FALSE,
  equal_scales = TRUE,
  normalize_log_loa = FALSE,
  keep_log_scale = FALSE
)
}
\arguments{
\item{ba_obj}{BA analysis object}

\item{show_subject_legend}{Show legend for subjects}

\item{equal_scales}{Plot the residuals on a plane with the scale of the original data.}

\item{keep_log_scale}{Show log transformed differences. If `FALSE` (default), values and parameters are exponentiated before plotting}
}
\description{
Creates a scatter plot, a standard Bland Altman plot and a residuals plot for assessing trending ability.
}
