% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{plot_BA_normalized_log}
\alias{plot_BA_normalized_log}
\title{Plot BA estimates from log transformed data on raw data.}
\usage{
plot_BA_normalized_log(
  ba_obj,
  show_subject_legend = FALSE,
  aspect_ratio = NULL
)
}
\arguments{
\item{ba_obj}{BA analysis object}

\item{show_subject_legend}{Show legend for subjects}

\item{aspect_ratio}{Set aspect ratio (x/y) between X and Y axis (sets \code{ggplot2::coord_fixed()}), Default (NULL) is automatic.}
}
\value{
Bland Altman style plot with relative differences plotted on absolute differences.
}
\description{
Plot BA estimates from log transformed data on raw data.
}
\examples{
plot_BA(compare_methods(CO, "ic", "rv", id_col = "sub", logtrans = TRUE))

}
