% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{plot_BA_residuals}
\alias{plot_BA_residuals}
\title{Plot intraindividual variation (model residuals) in differences and means.}
\usage{
plot_BA_residuals(
  ba_obj,
  show_subject_legend = FALSE,
  keep_original_scale = TRUE,
  normalize_log_loa = FALSE,
  exponentiate = FALSE,
  use_non_log_x_values = TRUE
)
}
\arguments{
\item{ba_obj}{BA analysis object}

\item{show_subject_legend}{Show legend for subjects}

\item{keep_original_scale}{Plot the residuals on a plane with the scale of the original data.}

\item{normalize_log_loa}{Plot using `plot_normalized_log_BA()`}

\item{exponentiate}{Exponentiate values and parameters before plotting}

\item{use_non_log_x_values}{Plot BA estimates from log transformed data on raw data.}
}
\description{
Plot intraindividual variation (model residuals) in differences and means.
}
