% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_measurement}
\alias{simulate_measurement}
\title{Simulate measurement}
\usage{
simulate_measurement(
  true_val,
  sub_id,
  residual_error_sd,
  mean_bias = 0,
  sub_bias_sd = 0,
  relative_errors = FALSE
)
}
\arguments{
\item{true_val}{A numeric vector of true values.}

\item{sub_id}{A factor or vector representing subject IDs. Must be the same length as var_true.}

\item{residual_error_sd}{Standard deviation of residual errors.}

\item{mean_bias}{Mean bias.}

\item{sub_bias_sd}{Standard deviation of subject-specific biases.}

\item{relative_errors}{Should errors be relative to true value.
If TRUE, errors are added on a log scale, and the result is exponentiated. To add a bias of 10\% (true_value * 1.1),
the mean_bias should be set to log(1.1). A SD of log(1.2) corresponds to ±20\%.}
}
\value{
A numeric vector representing simulated measurements.
}
\description{
Simulate measurement of a variable by adding random error.
The function allows adding a subject specific error (\code{sub_bias_sd}), that varies between subjects.
}
\examples{
df <- simulate_repeated_data(n_sub = 50, n_rep = 5, var_name = "true_co")
df$ref <- simulate_measurement(
  df$true_co,
  df$id,
  mean_bias = log(1.1),
  sub_bias_sd = log(1.2),
  residual_error_sd = log(1.1),
  relative_errors = TRUE
)
plot(df$true_co, df$ref, col=df$id)
abline(a=0, b=1)

}
