% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{add_confint}
\alias{add_confint}
\title{Add confidence intervals to BA analysis object.}
\usage{
add_confint(
  ba_obj,
  level = 0.95,
  nsim = 1999,
  .progress = "txt",
  PBargs = list(style = 3)
)
}
\arguments{
\item{ba_obj}{BA analysis object}

\item{level}{Confidence level (default is 0.95)}

\item{nsim}{Number of bootstrap samples}

\item{.progress, PBargs}{see \code{\link[lme4:bootMer]{lme4::bootMer()}}}
}
\value{
BA analysis object (\code{x}) with added confidence intervals
}
\description{
The confidence intervals are calculated with a percentile parametric bootstrap (see \code{\link[lme4:confint.merMod]{lme4::confint.merMod()}} for details).
}
