% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{as.data.frame.ba_analysis}
\alias{as.data.frame.ba_analysis}
\title{Convert Bland-Altman analysis object to data.frame}
\usage{
\method{as.data.frame}{ba_analysis}(x, ...)
}
\arguments{
\item{x}{Bland-Altman analysis object}

\item{...}{not used}
}
\description{
Convert Bland-Altman analysis object to data.frame
}
