% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{confint.ba_analysis}
\alias{confint.ba_analysis}
\title{Calculate confidence interval}
\usage{
\method{confint}{ba_analysis}(x, level = 0.95, nsim = 2000)
}
\description{
Calculate confidence interval
}
