% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stats.R
\name{confint.ba_analysis}
\alias{confint.ba_analysis}
\title{Calculate confidence interval}
\usage{
\method{confint}{ba_analysis}(ba_obj, level = 0.95, nsim = 2000, .progress = "txt")
}
\arguments{
\item{.progress}{}
}
\value{
Matrix of bootstrap confidence intervals for BA statistics.
}
\description{
Calculate confidence interval
}
