% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{plot_BA}
\alias{plot_BA}
\title{Create Bland Altman Plot}
\usage{
plot_BA(
  ba_obj,
  subject_legend = FALSE,
  normalize_log_loa = FALSE,
  exponentiate = FALSE,
  use_non_log_x_values = TRUE
)
}
\arguments{
\item{normalize_log_loa}{Plot BA estimates from log transformed data on raw data.}

\item{use_non_log_x_values}{}
}
\value{

}
\description{
Create Bland Altman Plot
}
\examples{

}
