% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{plot_BA}
\alias{plot_BA}
\title{Create Bland Altman Plot}
\usage{
plot_BA(
  ba_obj,
  fix_aspect_ratio = FALSE,
  show_subject_legend = FALSE,
  keep_log_scale = FALSE
)
}
\arguments{
\item{ba_obj}{BA analysis object}

\item{fix_aspect_ratio}{Use aspect ratio of 1 between X and Y axis (sets `coord_fixed()`)}

\item{show_subject_legend}{Show legend for subjects}

\item{keep_log_scale}{Show log transformed differences. If `FALSE` (default), values and parameters are exponentiated before plotting}
}
\value{
Bland Altman plot (ggplot)
}
\description{
Create Bland Altman Plot
}
\examples{
plot_BA(compare_methods(CO, "ic", "rv", id_col = "sub"))

}
