% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{plot_BA_combine}
\alias{plot_BA_combine}
\title{Plot all plots in extended Bland Altman analysis.}
\usage{
plot_BA_combine(
  ba_obj,
  aspect_ratio = NULL,
  show_subject_legend = FALSE,
  use_titles = TRUE,
  equal_scales = TRUE,
  keep_log_scale = FALSE,
  return_as_list = FALSE
)
}
\arguments{
\item{ba_obj}{BA analysis object}

\item{aspect_ratio}{Set aspect ratio (x/y) between X and Y axis (sets \code{ggplot2::coord_fixed()}), Default (NULL) is automatic.}

\item{show_subject_legend}{Show legend for subjects}

\item{use_titles}{Show default titles on plots.}

\item{equal_scales}{Plot the residuals on a plane with the scale of the original data.}

\item{keep_log_scale}{Show log transformed differences. If \code{FALSE} (default), values and parameters are exponentiated before plotting}

\item{return_as_list}{Return the three plots in a list. If \code{FALSE}, the plots are combined using \{patchwork\}.}
}
\description{
Creates a scatter plot, a standard Bland Altman plot and a residuals plot for assessing trending ability.
}
