% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{plot_BA_complete}
\alias{plot_BA_complete}
\title{Plot all plots in extended Bland Altman analysis.}
\usage{
plot_BA_complete(
  ba_obj,
  show_subject_legend = FALSE,
  equal_scales = TRUE,
  normalize_log_loa = FALSE,
  exponentiate = FALSE,
  use_non_log_x_values = TRUE
)
}
\arguments{
\item{ba_obj}{BA analysis object}

\item{show_subject_legend}{Show legend for subjects}

\item{equal_scales}{Plot the residuals on a plane with the scale of the original data.}

\item{normalize_log_loa}{Plot using `plot_normalized_log_BA()`}

\item{exponentiate}{Exponentiate values and parameters before plotting}

\item{use_non_log_x_values}{Plot BA estimates from log transformed data on raw data.}
}
\description{
Creates a scatter plot, a standard Bland Altman plot and a residuals plot for assessing trending ability.
}
