% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{plot_normalized_log_BA}
\alias{plot_normalized_log_BA}
\title{Plot BA estimates from log transformed data on raw data.}
\usage{
plot_normalized_log_BA(ba_obj, show_subject_legend = FALSE)
}
\arguments{
\item{show_subject_legend}{}
}
\description{
Plot BA estimates from log transformed data on raw data.
}
