% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ba_analysis}
\alias{print.ba_analysis}
\title{Print method for Bland-Altman analysis objects}
\usage{
\method{print}{ba_analysis}(x, ...)
}
\arguments{
\item{x}{Bland-Altman analysis object.}

\item{...}{not used.}
}
\description{
Print method for Bland-Altman analysis objects
}
